+++ CHANGELOG.md
@@ -15,6 +15,10 @@ All notable changes to this project will be documented in this file.
 
 
 
+
+## [6.4.1] - 2025-12-29
+### Fixed
+- NIIMBOT printer support added
 ## [6.4.0] - 2025-12-29
 ### Added
 ## [6.3.1] - 2025-12-29